requirements.txt
PyPDF2
python-docx
Pillow
pytesseract
google-generativeai
streamlit
pandas
xlsxwriter
